refactored app out of index for testing
